add user validation

+++ src/main/java/com/oluchi/restapi/EventRESR.java
@@ -7,6 +7,8 @@ import com.oluchi.restapi.database.RepositoryEvent;
 import com.oluchi.restapi.entity.Event;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,6 +16,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -30,6 +33,11 @@ public class EventRESR {
         return repository.findAll();
     }
 
+    @GetMapping
+    public List<Event> listByUserId(@RequestHeader String userId) {
+        return repository.findAllByUserId(userId);
+    }
+
     @GetMapping("{id}")
     public Optional<Event> findById(@PathVariable String id) {
 
@@ -37,18 +45,29 @@ public class EventRESR {
     }
 
     @PostMapping
-    public void saveEvent(@RequestBody Event event) {
+    public void saveEvent(@RequestBody Event event, @RequestHeader String userId) {
+        event.setUserId(userId);
         repository.save(event);
     }
 
     @PutMapping("{id}")
-    public void changeEvent(@RequestBody Event event, @PathVariable String id) {
+    public ResponseEntity changeEvent(@RequestBody Event event, @PathVariable String id, @RequestHeader String userId) {
+        Optional<Event> e = repository.findById(id);
+        if (!e.isPresent() || e.get().getUserId() != userId) {
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        }
         event.setId(id);
         repository.save(event);
+        return null;
     }
 
     @DeleteMapping("{id}")
-    public void deleteEvent(@PathVariable String id) {
+    public ResponseEntity deleteEvent(@PathVariable String id, @RequestHeader String userId) {
+        Optional<Event> e = repository.findById(id);
+        if (!e.isPresent() || e.get().getUserId() != userId) {
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        }
         repository.deleteById(id);
+        return null;
     }
 }
